+++ src/phantomfetch/engines/browser/__init__.py
@@ -1,9 +1,7 @@
 from .actions import actions_to_payload, execute_actions
-from .baas import BaaSEngine
 from .cdp import CDPEngine
 
 __all__ = [
-    "BaaSEngine",
     "CDPEngine",
     "actions_to_payload",
     "execute_actions",